fix: images and pages
