+++ onboarding-ui/pages/[market].tsx
@@ -2,13 +2,13 @@ import { GetStaticProps } from 'next'
 import { WithRouterProps } from 'next/dist/client/with-router'
 import Head from 'next/head'
 import * as React from 'react'
-import Auth from '../../../components/Auth'
-import { FunnelHeader } from '../../../components/FunnelHeader'
-import { PoolOnboarding } from '../../../components/Onboarding'
-import { PageContainer } from '../../../components/PageContainer'
-import { TinlakeProvider } from '../../../components/TinlakeProvider'
-import WithFooter from '../../../components/WithFooter'
-import config, { loadPoolsFromIPFS, Pool } from '../../../config'
+import Auth from '../components/Auth'
+import { FunnelHeader } from '../components/FunnelHeader'
+import { PoolOnboarding } from '../components/Onboarding'
+import { PageContainer } from '../components/PageContainer'
+import { TinlakeProvider } from '../components/TinlakeProvider'
+import WithFooter from '../components/WithFooter'
+import config, { loadPoolsFromIPFS, Pool } from '../config'
 
 interface Props extends WithRouterProps {
   market: 'rwa-market'
