+++ onboarding-ui/pages/pool/[root]/[slug]/index.tsx
@@ -2,14 +2,14 @@ import { GetStaticProps } from 'next'
 import Head from 'next/head'
 import { useRouter } from 'next/router'
 import * as React from 'react'
-import Auth from '../../../../../components/Auth'
-import { FunnelHeader } from '../../../../../components/FunnelHeader'
-import { IpfsPoolsProvider } from '../../../../../components/IpfsPoolsProvider'
-import { PoolOnboarding } from '../../../../../components/Onboarding'
-import { PageContainer } from '../../../../../components/PageContainer'
-import { TinlakeProvider } from '../../../../../components/TinlakeProvider'
-import WithFooter from '../../../../../components/WithFooter'
-import { IpfsPools, loadPoolsFromIPFS, Pool } from '../../../../../config'
+import Auth from '../../../../components/Auth'
+import { FunnelHeader } from '../../../../components/FunnelHeader'
+import { IpfsPoolsProvider } from '../../../../components/IpfsPoolsProvider'
+import { PoolOnboarding } from '../../../../components/Onboarding'
+import { PageContainer } from '../../../../components/PageContainer'
+import { TinlakeProvider } from '../../../../components/TinlakeProvider'
+import WithFooter from '../../../../components/WithFooter'
+import { IpfsPools, loadPoolsFromIPFS, Pool } from '../../../../config'
 
 interface Props {
   root: string
